+++ sdk/src/client/pouchdb.test.js
@@ -1,0 +1,11 @@
+import { describe, test } from "node:test";
+
+describe("pouchdb", () => {
+  describe("findLatestInteraction", () => {
+    // TODO
+  });
+
+  describe("saveInteraction", () => {
+    // TODO
+  });
+});
