+++ sdk/src/dal.test.js
@@ -1,14 +1,9 @@
 import { describe, test } from "node:test";
 import * as assert from "node:assert";
 
-import {
-  loadInteractionsWith,
-  loadTransactionDataWith,
-  loadTransactionMetaWith,
-} from "./dal.js";
+import { loadTransactionDataWith, loadTransactionMetaWith } from "./dal.js";
 
 const GATEWAY_URL = globalThis.GATEWAY || "https://arweave.net";
-const SEQUENCER_URL = globalThis.SEQUENCER_URL || "https://gw.warp.cc";
 const CONTRACT = "zc24Wpv_i6NNCEdxeKt7dcNrqL5w0hrShtSCcFGGL24";
 
 describe("dal", () => {
@@ -90,28 +85,6 @@ describe("dal", () => {
       assert.ok(result.json);
       assert.ok(result.text);
       assert.ok(result.ok);
-    });
-  });
-
-  describe("loadInteractions", () => {
-    test("load the interactions from the sequencer", async () => {
-      const CONTRACT = "SFKREVkacx7N64SIfAuNkMOPTocm42qbkKwzRJGfQHY";
-
-      const loadInteractions = loadInteractionsWith({
-        fetch,
-        SEQUENCER_URL,
-      });
-
-      const res = await loadInteractions({
-        id: CONTRACT,
-        from: "",
-        to: "",
-      }).toPromise();
-      assert.ok(res.length);
-      const [firstInteraction] = res;
-      assert.ok(firstInteraction.action);
-      assert.ok(firstInteraction.action.function);
-      assert.ok(firstInteraction.sortKey);
     });
   });
 });
